Modularize repository for easier customization
---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/tammyyang/data-mart-mockup?shareId=XXXX-XXXX-XXXX-XXXX).
