+++ index.html
@@ -50,6 +50,25 @@
         </div>
     </footer>
 
-    <script src="collections.js"></script>
+    <script type="module">
+        import { fetchData } from './modules/fetchData.js';
+        import { renderDatasets } from './modules/renderDatasets.js';
+        import { setupSearch } from './modules/setupSearch.js';
+
+        document.addEventListener('DOMContentLoaded', () => {
+            fetchData()
+                .then(data => {
+                    if (data) {
+                        renderDatasets(data);
+                        setupSearch(data, renderDatasets);
+                    }
+                })
+                .catch(error => {
+                    console.error('Error fetching collections:', error);
+                    document.getElementById('dataset-container').innerHTML = 
+                        '<div class="error">Failed to load datasets. Please try again later.</div>';
+                });
+        });
+    </script>
 </body>
 </html>
