+++ modules/getLocalImagePath.js
@@ -1,0 +1,4 @@
+export function getLocalImagePath(index, provider) {
+    const [firstName, lastName] = provider.split(' ');
+    return `images/${index + 1}.jpg`;
+}
